feat: Added new Participation role
