+++ src/Form/User/EditUserType.php
@@ -1,50 +1,52 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\User;
 
 use App\DataTransferObjects\UserAdminDTO;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class CreateUserType extends AbstractType
+class EditUserType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
+            ->add('id', HiddenType::class)
             ->add('username', TextType::class)
-            ->add('password', PasswordType::class)
+            ->add('password', PasswordType::class, [
+                'required' => false,
+            ])
             ->add('email', EmailType::class)
             ->add('roles', ChoiceType::class, [
                 'multiple' => true,
                 'expanded' => true,
                 'choices' => [
                     'User' => 'ROLE_USER',
+                    'Participant' => 'ROLE_PARTICIPANT',
                     'Admin' => 'ROLE_ADMIN',
                 ],
             ])
             ->add('credentialsExpired', CheckboxType::class, [
                 'required' => false,
-                'value' => true,
                 'label_attr' => [
                     'class' => 'checkbox-inline checkbox-switch',
                 ],
             ])
             ->add('isVerified', CheckboxType::class, [
                 'required' => false,
-                'value' => true,
                 'label_attr' => [
                     'class' => 'checkbox-inline checkbox-switch',
                 ],
             ])
             ->add('isPublic', CheckboxType::class, [
                 'required' => false,
-                'value' => false,
                 'label_attr' => [
                     'class' => 'checkbox-inline checkbox-switch',
                 ],
