+++ templates/admin/user/edit.html.twig
@@ -21,8 +21,35 @@
 
 {% block admin_title_actions %}
     <div class="btn-list">
-        <div class="btn-list">
-            <span class="d-none d-sm-inline">
+        {{ include('admin/user/_delete_form.html.twig') }}
+    </div>
+{% endblock %}
+
+{% block body %}
+    <div class="card">
+        {{ form_start(form) }}
+        <div class="card-header bg-body-secondary">
+            <h3 class="card-title">{{ "title.user_edit"|trans }}</h3>
+        </div>
+        <div class="card-body">
+            {{ form_row(form.username) }}
+
+            {{ form_row(form.password) }}
+
+            {{ form_row(form.email) }}
+
+            {{ form_row(form.roles) }}
+
+            <hr />
+
+            {{ form_row(form.credentialsExpired) }}
+
+            {{ form_row(form.isVerified) }}
+
+            {{ form_row(form.isPublic) }}
+        </div>
+        <div class="card-footer">
+            <div class="float-start">
                 <a class="btn btn-outline-secondary" href="{{ path('app_admin_user_index') }}">
                     <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-chevron-left" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
                         <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
@@ -30,52 +57,22 @@
                     </svg>
                     {{ "label.back_to_list"|trans }}
                 </a>
-            </span>
-        </div>
-    </div>
-{% endblock %}
-
-{% block body %}
-    <div class="container">
-        <div class="card">
-            <div class="card-header">
-                <h3 class="card-title">{{ "title.user_edit"|trans }}</h3>
             </div>
-            <div class="card-body">
-                {{ form_start(form) }}
-
-                {{ form_row(form.username) }}
-
-                {{ form_row(form.password) }}
-
-                {{ form_row(form.email) }}
-
-                {{ form_row(form.roles) }}
-
-                {{ form_row(form.credentialsExpired) }}
-
-                {{ form_row(form.isVerified) }}
-
-                {{ form_row(form.isPublic) }}
-            </div>
-            <div class="card-footer">
-                <div class="btn-list justify-content-end">
-                    <a href="{{ path('app_admin_user_show', {id: user.id}) }}" class="btn">
-                        {{ "label.cancel"|trans }}
-                    </a>
-                    <button type="submit" class="btn btn-primary" type="submit">
-                        <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-device-floppy" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
-                            <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
-                            <path d="M6 4h10l4 4v10a2 2 0 0 1 -2 2h-12a2 2 0 0 1 -2 -2v-12a2 2 0 0 1 2 -2" />
-                            <path d="M12 14m-2 0a2 2 0 1 0 4 0a2 2 0 1 0 -4 0" />
-                            <path d="M14 4l0 4l-6 0l0 -4" />
-                        </svg>
-                        {{ "label.save_changes"|trans }}
-                    </button>
-                </div>
-
-                {{ form_end(form) }}
+            <div class="btn-list justify-content-end">
+                <a href="{{ path('app_admin_user_show', {id: user.id}) }}" class="btn">
+                    {{ "label.cancel"|trans }}
+                </a>
+                <button type="submit" class="btn btn-primary" type="submit">
+                    <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-device-floppy" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                        <path d="M6 4h10l4 4v10a2 2 0 0 1 -2 2h-12a2 2 0 0 1 -2 -2v-12a2 2 0 0 1 2 -2" />
+                        <path d="M12 14m-2 0a2 2 0 1 0 4 0a2 2 0 1 0 -4 0" />
+                        <path d="M14 4l0 4l-6 0l0 -4" />
+                    </svg>
+                    {{ "label.save_changes"|trans }}
+                </button>
             </div>
         </div>
+        {{ form_end(form) }}
     </div>
 {% endblock %}
