+++ templates/admin/user/new.html.twig
@@ -19,29 +19,15 @@
 {% endblock %}
 
 {% block admin_title_actions %}
-    <div class="btn-list">
-        <div class="btn-list">
-            <span class="d-none d-sm-inline">
-                <a class="btn btn-outline-secondary" href="{{ path('app_admin_user_index') }}">
-                    <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-chevron-left" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
-                        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
-                        <path d="M15 6l-6 6l6 6" />
-                    </svg>
-                    {{ "label.back_to_list"|trans }}
-                </a>
-            </span>
-        </div>
-    </div>
 {% endblock %}
 
 {% block body %}
-    <div class="container">
-        <div class="card">
-            <div class="card-header">
+    <div class="card">
+        {{ form_start(form) }}
+            <div class="card-header bg-body-secondary">
                 <h3 class="card-title">{{ "title.user_create"|trans }}</h3>
             </div>
             <div class="card-body">
-                {{ form_start(form) }}
 
                 {{ form_row(form.username) }}
 
@@ -51,6 +37,8 @@
 
                 {{ form_row(form.roles) }}
 
+                <hr />
+
                 {{ form_row(form.credentialsExpired) }}
 
                 {{ form_row(form.isVerified) }}
@@ -58,6 +46,17 @@
                 {{ form_row(form.isPublic) }}
             </div>
             <div class="card-footer">
+                <div class="float-start">
+                    <span class="d-none d-sm-inline">
+                        <a class="btn btn-outline-secondary" href="{{ path('app_admin_user_index') }}">
+                            <svg xmlns="http://www.w3.org/2000/svg" class="icon icon-tabler icon-tabler-chevron-left" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
+                                <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+                                <path d="M15 6l-6 6l6 6" />
+                            </svg>
+                            {{ "label.back_to_list"|trans }}
+                        </a>
+                    </span>
+                </div>
                 <div class="btn-list justify-content-end">
                     <a href="{{ path('app_admin_user_index') }}" class="btn">
                         {{ "label.cancel"|trans }}
@@ -72,9 +71,7 @@
                         {{ "label.create_account"|trans }}
                     </button>
                 </div>
-
-                {{ form_end(form) }}
             </div>
-        </div>
+        {{ form_end(form) }}
     </div>
 {% endblock %}
